pyproject.toml
[project]
name = "smart_dictation"
version = "0.1.0"
description = "Fast dictation app based on OpenAI Whisper with support for LLM transformations"
readme = "README.md"
authors = [
    { name = "Piotr Czapla", email = "piotr.czapla@n-waves.com" },
]
requires-python = ">=3.10,<3.12"

dependencies = [
    "pyaudio>=0.2.13,<1.0.0",
    "pynput>=1.7.6,<2.0.0",
    "pydantic>=2.9.2",
    "pyperclip>=1.9.0",
    "pydantic-settings>=2.5.2",
    "structlog>=24.4.0",
    "numpy>=2.1.2,<3",
    "pywhispercpp @ git+https://github.com/abdeladim-s/pywhispercpp.git@41545c8638f3a444bab26509738a70e3b02d18b2",
    "pyobjc-core>=9.2; platform_system=='Darwin'",
    "pyobjc-framework-Quartz>=9.2; platform_system=='Darwin'",
    "pyobjc-framework-AppKit>=9.2; platform_system=='Darwin'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3,<9",
    "pytest-asyncio>=0.24.0,<0.25",
]

[tool.setuptools]
packages = {find = {where = ["src"], exclude = ["src.experiments*"]}}
package-dir = {"" = "src"}

[project.scripts]
smart_dictation = "smart_dictation:app.main"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

